+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-native-tableview",
   "version": "1.7.0",
   "description": "Native iOS TableView wrapper for React Native",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -31,6 +31,7 @@
     "prop-types": "^15.6.0"
   },
   "devDependencies": {
+    "prettier": "^1.8.2",
     "react": "16.0.0",
     "react-native": "^0.50.3"
   }
